requirements.txt
Django==5.0.6
sqlparse==0.5.0
psycopg2-binary>=2.9.6
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
djangorestframework-jwt==1.11.0
django-cors-headers==4.4.0
django-rest-authtoken==2.1.4
requests==2.32.3
tqdm==4.66.4

